Feature/currency

+++ OnlinerTracker/OnlinerTracker.BusinessLogic/Models/Price.cs
@@ -5,9 +5,9 @@ namespace OnlinerTracker.BusinessLogic.Models
 	public class Price
 	{
 		[JsonProperty("min")]
-		public int Min { get; set; }
+		public decimal Min { get; set; }
 
 		[JsonProperty("max")]
-		public int Max { get; set; }
+		public decimal Max { get; set; }
 	}
 }
